Correcao de bug no controller de Edital, Correcao de bugs no Controller de Programas, Adicionado verificacao se ha algum edital vinculado ao programa antes de exculi-lo

+++ app/Http/Controllers/ProgramaController.php
@@ -13,7 +13,7 @@ use App\Models\Edital;
 use App\Models\Orientador;
 use App\Models\Disciplina;
 use App\Http\Controllers\EditalController;
-
+use Exception;
 
 class ProgramaController extends Controller
 {
@@ -72,13 +72,12 @@ class ProgramaController extends Controller
             $programa = new Programa();
             $programa->nome = $request->nome;
             $programa->descricao = $request->descricao;
-            $programa->valor_bolsa = $request->valor_bolsa;
             $programa->save();
 
-            foreach($request->servidors as $id_servidor){
+            if ($request->servidor){
                 $programa_servidor = new Programa_servidor();
                 $programa_servidor->programa_id = $programa->id;
-                $programa_servidor->id_servidor = $id_servidor;
+                $programa_servidor->servidor_id = $request->servidor;
                 $programa_servidor->save();
             }
 
@@ -116,7 +115,7 @@ class ProgramaController extends Controller
         $idsservidorsDoPrograma = [];
 
         foreach ($programa->programa_servidors as $programa_servidor){
-            $idsservidorsDoPrograma[] = $programa_servidor->id_servidor;
+            $idsservidorsDoPrograma[] = $programa_servidor->servidor_id;
         }
 
         return view("Programa.editar", compact('programa', 'servidors', 'idsservidorsDoPrograma'));
@@ -130,24 +129,19 @@ class ProgramaController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(ProgramaUpdateFormRequest $request, $id)
-    {
+    {   
         DB::beginTransaction();
         try{
             $programa = Programa::find($id);
             $programa->nome = $request->nome ? $request->nome : $programa->nome;
             $programa->descricao = $request->descricao ? $request->descricao : $programa->descricao;
-            $programa->valor_bolsa = $request->valor_bolsa ? $request->valor_bolsa : $programa->valor_bolsa;
             $programa->update();
 
-            Programa_servidor::where("programa_id", $programa->id)->delete();
-
-            if ($request->servidors){
-                foreach($request->servidors as $id_servidor){
-                    $programa_servidor = new Programa_servidor();
+            if ($request->servidor){
+                    $programa_servidor = Programa_servidor::where("programa_id", $programa->id)->first();
                     $programa_servidor->programa_id = $programa->id;
-                    $programa_servidor->id_servidor = $id_servidor;
-                    $programa_servidor->save();
-                }
+                    $programa_servidor->servidor_id = $request->servidor;
+                    $programa_servidor->update();
             }
 
             DB::commit();
@@ -156,7 +150,7 @@ class ProgramaController extends Controller
 
         } catch(exception $e){
             DB::rollback();
-            return redirect()->back()->withErrors( "Falha ao editar Programa. tente novamente mais tarde." );
+            return redirect()->back()->withErrors( "Falha ao editar Programa. Tente novamente mais tarde." );
         }
     }
 
@@ -169,22 +163,30 @@ class ProgramaController extends Controller
 
     public function destroy(Request $request)
     {
-        DB::beginTransaction();
-        try{
-            $id = $request->only(['id']);
-            $programa = Programa::findOrFail($id)->first();
+        $edital = Edital::where("programa_id", $request->id)->first();
+        if ($edital)
+        {
+            return redirect()->back()->withErrors( "Falha ao deletar Programa. Existem editais vinculados a ele." );
+        } else{
+        
+	
+            DB::beginTransaction();
+            try{
+                $id = $request->only(['id']);
+                $programa = Programa::findOrFail($id)->first();
 
-            Programa_servidor::where("programa_id", $programa->id)->delete();
+                Programa_servidor::where("programa_id", $programa->id)->delete();
 
-            $programa->delete();
+                $programa->delete();
 
 
-            DB::commit();
+                DB::commit();
 
-            return redirect('/programas')->with('sucesso', 'Programa deletado com sucesso.');
-        } catch(exception $e){
-            DB::rollback();
-            return redirect()->back()->withErrors( "Falha ao deletarm um Programa. tente novamente mais tarde." );
+                return redirect('/programas')->with('sucesso', 'Programa deletado com sucesso.');
+            } catch(exception $e){
+                DB::rollback();
+                return redirect()->back()->withErrors( "Falha ao deletarm um Programa. tente novamente mais tarde." );
+            }
         }
     }
